Run oim-panels in the OIM pipeline; retire the template-match truth path (#290)
`pipeline_oim` still ran the `oim-split-truth` step, whose input is
oim/pN__i.jpg crops downloaded by `download_oim_splits` -- a module that is not
and never has been a registered subcommand (`git log -S` on cli.py finds no
commit adding it). Missing crops raise ValueError, which that command catches
and turns into "Warning: skipping <page>", so a new OIM volume ran a step that
could only warn-and-skip and never ran the replacement. Richmond, the newest
volume, has zero crops. This was a silent no-split-truth bug, not just dead
code, and the pipeline now runs `mapsnap oim-panels`.

Deleting the old path is safe by measurement, not assumption: all 122 split
parents across the 18 truth volumes have an oim/pN.panels.json, and every one
of the 122 records "source": "oim-region-boundaries" -- #274's reader. The
template-match fallback #273 wanted to keep has no users.

Removed: download_oim_splits.py, oim_truth.py, their tests, the
`oim-split-truth` CLI entry, and scripts/make_panel_truth.py (superseded by
oim-panels; the #83 harness scores against oim/pN.panels.json directly, and
nothing referenced the script but its own docstring).

fix_truth_splits stays -- it is registered, and compare imports
gcp_containment from it -- so its one remaining dependency, panel_polygon, is
inlined there with a note that #273 disproved this technique for BUILDING
truth. It survives only to re-derive a crop offset by subtracting the polygon
from a stored ring, where small shape errors cancel.

compare's "run this to fix it" advice pointed at the retired command; it names
oim-panels now, along with the test asserting that message.
