+++ mapsnap/compare_iiif_georef.py
@@ -833,12 +833,12 @@ def compare_pages(
             # be matched to its truth splits — every one scores as unplaced,
             # silently understating the volume (KC sat ~4 net points low this
             # way). Loud warning rather than assertion: the fix is a data run
-            # (`mapsnap oim-split-truth`), and comparison of the other pages
+            # (`mapsnap oim-panels`), and comparison of the other pages
             # is still meaningful in the meantime.
             print(
                 f"Warning: {page_key} has {len(truth_splits)} truth split(s) but "
                 f"{oim_dir / f'{page_key}.panels.json'} is missing; its placements "
-                "will score as unplaced. Run `mapsnap oim-split-truth` (see its "
+                "will score as unplaced. Run `mapsnap oim-panels` (see its "
                 "docstring for the required raw/ and oim/ images).",
                 file=sys.stderr,
             )
